fix np.float64 doctest update

+++ leap/utils.py
@@ -168,9 +168,9 @@ def compute_ordinal_regression(
     Examples:
 
         >>> compute_ordinal_regression(0, 0.5)
-        [np.float64(0.6224593312018546), np.float64(0.3775406687981454)]
+        [0.6224593312018546, 0.3775406687981454]
         >>> compute_ordinal_regression(0, [0.5, 1.5])
-        [np.float64(0.6224593312018546), np.float64(0.19511514499178906), np.float64(0.18242552380635635)]
+        [0.6224593312018546, 0.19511514499178906, 0.18242552380635635]
     """
 
     if isinstance(θ, float):
